Update from RP before fixing the reading problems

+++ lekrws1rrp3Assignment2/src/lekrws1rrp3Assignment2/Competitordetails.java
@@ -2,6 +2,6 @@ package lekrws1rrp3Assignment2;
 
 public interface Competitordetails {
 	
-	 public abstract int getOverallScore();
+	 public abstract double getOverallScore();
 	 public abstract String getfullDetails();
 }
